requirements.txt
python-dotenv==1.0.1
gunicorn==23.0.0
requests==2.31.0
boto3==1.34.162
Flask==3.0.3
flask_cors==5.0.0
langchain==0.1.20
langchain-openai==0.1.7
langchain-google-genai==1.0.4
flashrank[listwise]==0.2.9
chromadb==0.5.23
pandas==2.2.3